+++ tops_backend/src/memory/device/device_box.rs
@@ -471,7 +471,7 @@ impl<T: DeviceCopy> Pointer for TopsDeviceBox<T> {
 // impl<T: DeviceCopy> crate::memory::private::Sealed for DeviceBox<T> {}
 impl<T: DeviceCopy> CopyDestination<T> for TopsDeviceBox<T> {
     fn copy_from(&mut self, val: &T) -> DeviceResult<()> {
-        let size = mem::size_of::<T>();
+        let size = mem::size_of::<T>() as u64;
         if size != 0 {
             unsafe {
                 driv::topsMemcpyHtoD(self.ptr.as_raw(), val as *const T as *mut c_void, size)
@@ -482,7 +482,7 @@ impl<T: DeviceCopy> CopyDestination<T> for TopsDeviceBox<T> {
     }
 
     fn copy_to(&self, val: &mut T) -> DeviceResult<()> {
-        let size = mem::size_of::<T>();
+        let size = mem::size_of::<T>() as u64;
         if size != 0 {
             unsafe {
                 driv::topsMemcpyDtoH(
@@ -499,7 +499,7 @@ impl<T: DeviceCopy> CopyDestination<T> for TopsDeviceBox<T> {
 
 impl<T: DeviceCopy> CopyDestination<TopsDeviceBox<T>> for TopsDeviceBox<T> {
     fn copy_from(&mut self, val: &TopsDeviceBox<T>) -> DeviceResult<()> {
-        let size = mem::size_of::<T>();
+        let size = mem::size_of::<T>() as u64;
         if size != 0 {
             unsafe { driv::topsMemcpyDtoD(self.ptr.as_raw(), val.ptr.as_raw(), size).to_result()? }
         }
@@ -507,7 +507,7 @@ impl<T: DeviceCopy> CopyDestination<TopsDeviceBox<T>> for TopsDeviceBox<T> {
     }
 
     fn copy_to(&self, val: &mut TopsDeviceBox<T>) -> DeviceResult<()> {
-        let size = mem::size_of::<T>();
+        let size = mem::size_of::<T>() as u64;
         if size != 0 {
             unsafe { driv::topsMemcpyDtoD(val.ptr.as_raw(), self.ptr.as_raw(), size).to_result()? }
         }
@@ -518,7 +518,7 @@ impl<T: DeviceCopy> CopyDestination<TopsDeviceBox<T>> for TopsDeviceBox<T> {
 impl<T: DeviceCopy> AsyncCopyDestination<T> for TopsDeviceBox<T> {
     type StreamT = TopsStream;
     unsafe fn async_copy_from(&mut self, val: &T, stream: &Self::StreamT) -> DeviceResult<()> {
-        let size = mem::size_of::<T>();
+        let size = mem::size_of::<T>() as u64;
         if size != 0 {
             driv::topsMemcpyHtoDAsync(
                 self.ptr.as_raw(),
@@ -532,7 +532,7 @@ impl<T: DeviceCopy> AsyncCopyDestination<T> for TopsDeviceBox<T> {
     }
 
     unsafe fn async_copy_to(&self, val: &mut T, stream: &Self::StreamT) -> DeviceResult<()> {
-        let size = mem::size_of::<T>();
+        let size = mem::size_of::<T>() as u64;
         if size != 0 {
             driv::topsMemcpyDtoHAsync(
                 val as *mut _ as *mut c_void,
@@ -549,7 +549,7 @@ impl<T: DeviceCopy> AsyncCopyDestination<T> for TopsDeviceBox<T> {
 impl<T: DeviceCopy> AsyncCopyDestination<TopsDeviceBox<T>> for TopsDeviceBox<T> {
     type StreamT = TopsStream;
     unsafe fn async_copy_from(&mut self, val: &TopsDeviceBox<T>, stream: &Self::StreamT) -> DeviceResult<()> {
-        let size = mem::size_of::<T>();
+        let size = mem::size_of::<T>() as u64;
         if size != 0 {
             driv::topsMemcpyDtoDAsync(self.ptr.as_raw(), val.ptr.as_raw(), size, stream.as_inner())
                 .to_result()?
@@ -558,7 +558,7 @@ impl<T: DeviceCopy> AsyncCopyDestination<TopsDeviceBox<T>> for TopsDeviceBox<T> 
     }
 
     unsafe fn async_copy_to(&self, val: &mut TopsDeviceBox<T>, stream: &Self::StreamT) -> DeviceResult<()> {
-        let size = mem::size_of::<T>();
+        let size = mem::size_of::<T>() as u64;
         if size != 0 {
             driv::topsMemcpyDtoDAsync(val.ptr.as_raw(), self.ptr.as_raw(), size, stream.as_inner())
                 .to_result()?
